Implement type encoding of integers
Implement the encoding for major types `0` (positive integer) and
`1` (negative integer) [2.1]. The core var uint encoding function has
been implemented as a seperate template because it'll be reused to
implement most of the encoding functions of the remaining major types.

The `type_encoder` is only meant for encoding the major type and the
additional information field and not as the public encoding API. Even
though the latter will be implemented in terms of the type_encoder. The
difference between these APIs will manifest itself for the string, map,
etc. types where the `type_encoder` will only encode the length, but not
the actual content.

The integer encoding unit tests currently only cover all edge cases for
the positive value range. The negative value range currently lacks tests
for values such as `std::int16_t{-1 - 0xff}` which will be encoded as
one value byte (and not two as the two's complement representation would
require). Thus the n byte lower bound test cases are missing.
Furthermore the examples from Appendix A have _not_ been added yet.