Add the basic_encoder customization point
Add the basic_encoder class template which serves as a type dispatcher.
basic_encoder is designed as a functor which captures the output stream
in its instance and only takes the value to be encoded as operator()
parameter. This allows one to pass basic_encoder instances directly to
most stdlib container algorithms.

basic_encoder is expected to be specialized for user types, but is not
the actual surface encoding API because the design prevents any template
argument deduction of the stream type.

Add specializations for the supported fundamental types which just
forward their arguments to the appropriate type_encoder function.