Refactor encode API stream passing
Pass the output stream as a normal function parameter instead of
carrying it as functor state. This drastically reduces the boilerplate
for basic_encoder specializations. The stream template parameter however
is still specified in the basic_encoder class template in order to allow
the encodeable constraints to be specified in terms of the concrete
basic_encoder invocation.