Allow `(object|tuple)_def` to be versioned
Add a `std::uint32_t version` field to `(object|tuple)_def` for the
descriptor version. Reserve `0xffff'ffff` as
`dplx::dp::null_def_version` which represents unversioned descriptors
and CBOR (map|array) items. Add a flag to `(object|tuple)_def` which
can be used to opt-in to the default decoder implementation for
versioned descriptors.