Implement memory IO streams on top of a buffer
Add  a buffer view abstraction which tracks a memory allocation and a
shifting unprocessed/unconsumed area at the end of said allocation. The
buffer has been designed to efficiently update, read from or write to
the unprocessed area which makes it particularly suitable for memory IO
streams.

Implement bounded memory input and output stream operations on top of
the buffer view, i.e. a buffer view can be fed to both encode and decode
functions.