Implement chunking stream CRTP helpers
Add an abstraction for streams which resemble (lazy) scatter-gather
buffers.

The input stream reads need an intermediate buffer for "fusing" the
buffers together, because deeppack stream reads need to be linearized
within memory.

The output stream writes need to be linearized similiarly, but can't use
a lazy buffer like the input streams. A stream wise buffer would need a
flush operation, therefore the write buffer is embedded within the
write_proxy and only used if the write wouldn't fit in the current chunk.