Allow computed member access for tuple|object defs
Add `property_fun` and `tuple_member_fun` which take a callable type
instead of a member object pointer which is called to obtain a pointer
to the member to be decoded or encoded. The callable type must
explicitly specify its `class_type` and `value_type` in order to allow
its `operator()` to be a template which only needs to be instantiatable
when the (de|en)coder is called.

- Provide `member_accesor_base` template to reduce the boilerplate.