Replace workaround selection macros
Add the `DPLX_DP_WORKAROUND` macro which enables in code workarounds for
older but still supported (compiler) versions.

Add the `DPLX_DP_WORKAROUND_TESTED_AT` macro which ought to be used for
in code workarounds where the required bugfix has not yet been
associated with an available version. Therefore we provide the
`DPLX_DP_FLAG_OUTDATED_WORKAROUNDS` configuration option which maps all
occurences which should be checked again for the current component
version to compiler errors. E.g. if a workaround was last validated for
version 1.0.0 and we now build with version 1.0.1, we would emit a
compiler error if (and only if) `DPLX_DP_FLAG_OUTDATED_WORKAROUNDS=1`.

- Replace all `BOOST_PREDEF_WORKAROUND` usages with one of the aboves.
- Add a config header with cmake pass through options.