Add string emit and encode APIs
Add core emit APIs for emitting strings _with_ content and not just the
prefix. These APIs take a pointer and size for efficiency reasons.

Add `codec::encode` support for stdlib strings and string_views.
However, we only support `char8_t` and `char` due to the fact that there
is no transcoding support in the stdlib.