Implement a stream for discarding output
The `encoded_size_of()` functions may depend on encoder settings and
therefore need to take an `emit_context` to be correctly implementable.
Since the `emit_context` requires a stream we need a simple in-memory
discarding one.

As a side benefit this would allow for implementing the `size_of()`
function in terms of the encoder.