Add test utils for output streams and emit contexts
Add a `simple_test_output_stream` which is a fixed size memory stream
with the option to delay buffer provisioning. Add a `test_output_stream`
which works similarily while additionally supporting multiple disjunct
buffers instead of a single contiguous one.

Add `emit_context` wrappers which fuse one of the aforementioned streams
with an `emit_context` due to the fact that codec tests need to
instantiate the emit_contexts themselves.