Merge `encoded_size_of` constraint into `encodable`
It is good practice to always provide a `size_of` implementation with an
`encode` function, i.e. making it mandatory shouldn't be a big issue.
If it were to be split of again one would need to find a good name, too.
The old `has_codec_size_of` isn't nice enough to become a public API.
And last but not least, it is always easier to relax the encodable
concept than to tighten it…