feat!: Pass emit_context by (non-const) reference
In order to allow storing state in the emit_context in a future API
extension it needs to be mutable. Note that since it stores the
output_buffer by reference, there are no concerns w.r.t. codecs
replacing the current output_buffer.

BREAKING CHANGE: The `(de|en)code()` APIs won't accept an
		 `emit_context const &` argument anymore