feat: Add a non-owning memory input stream
Add the most simple `input_buffer` implementation wrapping an existing
memory buffer. Allow rvalue references of byte spans to be converted to
said buffer in order to allow quick decoding of arbitrary buffers.
lvalue byte span references are explicitly forbidden, due to the fact
that there is no way to relay consumed byte information to the caller
in a ergonomic way. E.g. updating the span object to the remaining bytes
leaves the owner without a way to recover the original span and updating
const references is outright impossible.